# openapi/main.yaml
openapi: 3.0.0
info:
  title: Pooper API
  version: 1.0.0
  description: |
    API for the Pooper chat application.
    Provides endpoints for chat, MCP, tools, analytics, realtime, config, prompt, and preview services.
servers:
  - url: http://localhost:4100
    description: Local development server
...
components:
  responses:
    ApiError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    ApiError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Tool:
      type: object
      required:
        - id
        - name
        - serverId
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        serverId:
          type: string
          format: uuid
        type:
          type: string
          enum: [system, plugin, custom]
          default: custom
        parameters:
          type: array
          items:
            type: object
            required:
              - name
              - type
            properties:
              name:
                type: string
              type:
                type: string
              description:
                type: string
              required:
                type: boolean
                default: false
              schema:
                type: object
                additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
paths:
...
  /api/config:
    get:
      tags:
        - Config
      summary: Get configuration
      description: Retrieve application configuration
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    put:
      tags:
        - Config
      summary: Update configuration
      description: Update application configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

  # MCP routes
  /api/mcp/servers:
    get:
      tags:
        - MCP
      summary: Get MCP servers
      description: Retrieve list of MCP servers
      responses:
        '200':
          description: MCP servers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '500':
          $ref: '#/components/responses/ApiError'
    post:
      tags:
        - MCP
      summary: Create MCP server
      description: Create a new MCP server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
                - url
              properties:
                name:
                  type: string
                type:
                  type: string
                url:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: MCP server created successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
  /api/mcp/servers/{serverId}:
    get:
      tags:
        - MCP
      summary: Get MCP server details
      description: Retrieve details for a specific MCP server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: MCP server details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
    
    delete:
      tags:
        - MCP
      summary: Delete MCP server
      description: Delete a specific MCP server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: MCP server deleted successfully
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
  /api/mcp/tools:
    get:
      tags:
        - MCP
      summary: Get MCP tools
      description: Retrieve list of MCP tools
      responses:
        '200':
          description: MCP tools retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
        '500':
          $ref: '#/components/responses/ApiError'
    post:
      tags:
        - MCP
      summary: Create MCP tool
      description: Create a new MCP tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tool'
      responses:
        '201':
          description: MCP tool created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
  /api/mcp/tools/{toolId}:
    get:
      tags:
        - MCP
      summary: Get MCP tool
      description: Get a specific MCP tool by ID
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: MCP tool retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
    
    put:
      tags:
        - MCP
      summary: Update MCP tool
      description: Update a specific MCP tool
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tool'
      responses:
        '200':
          description: MCP tool updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
    
    delete:
      tags:
        - MCP
      summary: Delete MCP tool
      description: Delete a specific MCP tool
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: MCP tool deleted successfully
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
  /api/mcp/tools/{name}/execute:
    post:
      tags:
        - MCP
      summary: Execute MCP tool
      description: Execute a specific MCP tool
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                params:
                  type: object
                  additionalProperties: true
                serverId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Tool executed successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
...
  /api/mcp/config:
    get:
      tags:
        - MCP
      summary: Get MCP configuration
      description: Retrieve MCP configuration
      responses:
        '200':
          description: MCP configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
    put:
      tags:
        - MCP
      summary: Update MCP configuration
      description: Update MCP configuration
      parameters:
        - name: createBackup
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Whether to create a backup of the current configuration before updating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: MCP configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  backupCreated:
                    type: boolean
                  backupPath:
                    type: string
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
  /api/mcp/config/backup:
    get:
      tags:
        - MCP
      summary: Get configuration backups
      description: Get a list of MCP configuration backups
      responses:
        '200':
          description: Configuration backups retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  backups:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        path:
                          type: string
        '500':
          $ref: '#/components/responses/ApiError'
    
    post:
      tags:
        - MCP
      summary: Create backup
      description: Create a backup of the current MCP configuration
      responses:
        '200':
          description: Configuration backup created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  backupPath:
                    type: string
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
  /api/mcp/config/backup/{backupId}:
    get:
      tags:
        - MCP
      summary: Get backup
      description: Get a specific MCP configuration backup
      parameters:
        - name: backupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration backup retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
    
    post:
      tags:
        - MCP
      summary: Restore backup
      description: Restore a specific MCP configuration backup
      parameters:
        - name: backupId
          in: path
          required: true
          schema:
            type: string
        - name: createBackup
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Whether to create a backup of the current configuration before restoring
      responses:
        '200':
          description: Configuration backup restored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  backupCreated:
                    type: boolean
                  backupPath:
                    type: string
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
    
    delete:
      tags:
        - MCP
      summary: Delete backup
      description: Delete a specific MCP configuration backup
      parameters:
        - name: backupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration backup deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
  /api/mcp/config/validate:
    post:
      tags:
        - MCP
      summary: Validate configuration
      description: Validate an MCP configuration without saving it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Configuration validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        message:
                          type: string
                        severity:
                          type: string
                          enum: [error, warning, info]
        '500':
          $ref: '#/components/responses/ApiError'

  # Prompt routes
...